Allows adding content via file upload in instancewebapp UI
git-svn-id: https://svn.duraspace.org/duracloud-priv@49 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ trunk/storageprovider/src/main/java/org/duraspace/storage/StorageProvider.java
@@ -127,7 +127,7 @@ public interface StorageProvider {
     public void addContent(String spaceId,
                            String contentId,
                            String contentMimeType,
-                           int contentSize,
+                           long contentSize,
                            InputStream content)
     throws StorageException;
